requirements.txt
discord.py==0.16.12
discord.py[voice]~=0.16.12
pip
youtube_dl
colorlog
asyncio
requests